requirements.txt
tqdm
numpy
pyyaml
librosa
mir_eval
matplotlib
tensorboardX
Pillow
PyTorch